Fix: fixed Login method returns and properly created payload struct
