+++ cmd/insprd/auth/payload.go
@@ -1,0 +1,10 @@
+package auth
+
+//Payload is information caried by a Inspr acceess token
+type Payload struct {
+	UID        string
+	Role       int
+	Scope      []string
+	Refresh    string
+	RefreshURL string
+}
