+++ cmd/uid_provider/client/interface.go
@@ -1,14 +1,10 @@
 package client
 
-import "context"
+import (
+	"context"
 
-type Payload struct {
-	UID        string
-	Role       int
-	Scope      []string
-	Refresh    string
-	RefreshURL string
-}
+	"gitlab.inspr.dev/inspr/core/cmd/insprd/auth"
+)
 
 type User struct {
 	UID      string
@@ -18,13 +14,9 @@ type User struct {
 }
 
 type RedisManager interface {
-	// creates payload and sends it to insprd
-	// when creating the payload, generetes the Refresh Token (cryptografado)
-	Login(ctx context.Context, uid, pwd string) error // asks Insprd to generate token and saves it into file
-
+	Login(ctx context.Context, uid, pwd string) (string, error)
 	CreateUser(ctx context.Context, uid string, newUser User) error
 	DeleteUser(ctx context.Context, uid, usrToBeDeleted string) error
 	UpdatePassword(ctx context.Context, uid, usrToBeUpdated, newPwd string) error
-
-	RefreshToken(ctx context.Context, refreshToken string) (Payload, error)
+	RefreshToken(ctx context.Context, refreshToken string) (auth.Payload, error)
 }
